WIP : build more gRPC cloud APIs by default
the goal is to build google-cloud-cpp as it was a year or so ago. with all the cloud APIs
not just the storage, bigtable ones.

- for now:
   - cloud datastore v1
   - cloud logging v2
   - cloud monitoring v3
   - cloud devtools cloudtracing v2
   - cloud devtools errorreporting v1beta1
   - cloud pubsub v1
   - cloud watcher v1
   - supporting api, rpc protos

in addition to bigtable, iam v1, needed to support

- but:
   - no firestore v1 gRPC API.
   - no cloud ML, auto ML, vision, speech ...
   - no cloud spanner
   - no cloud function.
   - no ads, home,

in addition, am building as shared libs, but not checking that the libraries have no undefined symbols.